organize wiiuse source a little
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@3743 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Externals/WiiUseSrc/Src/ir.cpp
@@ -663,7 +663,7 @@ static float ir_distance(struct ir_dot_t* dot) {
 	xd = dot[i2].x - dot[i1].x;
 	yd = dot[i2].y - dot[i1].y;
 
-	return sqrt(xd*xd + yd*yd);
+	return sqrt(float(xd*xd + yd*yd));
 }
